skill szepites, szerver jog

+++ app/controllers/skills_controller.rb
@@ -1,4 +1,6 @@
 class SkillsController < ApplicationController
+  before_action :authenticate_user!
+  
   before_action :set_skill, only: [:destroy, :confirm]
   before_action :set_user, only: [:create, :confirm]
   skip_before_filter :verify_authenticity_token
@@ -7,6 +9,7 @@ class SkillsController < ApplicationController
   # GET /skills.json
   def index
     if params[:user_id]
+      
       @skills=User.find(params[:user_id]).users_skills.map{|s| s.formatted}
     else
       @skills = Skill.all
@@ -15,22 +18,32 @@ class SkillsController < ApplicationController
   end
 
   def create
+    #TODO: getto jogkor kezeles
+    raise CanCan::AccessDenied if !current_user.has_role?(:godfather) || !(current_user.id!=params[:user_id])
+    
     @skill = Skill.find_or_create_by(name: params[:skill][:name])
-    @users_skill=@user.users_skills.create(skill: @skill)
+    @users_skill=@user.users_skills.create(skill: @skill) if @user.users_skills.where(user: @user, skill: @skill).count==0
     render json: @users_skill.formatted
   end
 
   def destroy
     if params[:user_id]
+      raise CanCan::AccessDenied if !current_user.has_role?(:godfather) || !(current_user.id!=params[:user_id])
+      
       UsersSkill.where(user_id: params[:user_id], skill_id: params[:id]).destroy_all
     else
+      raise CanCan::AccessDenied if !current_user.has_role?(:admin)
+      
       @skill.destroy
     end
     render json: {status: :ok}
   end
   
   def confirm
+    raise CanCan::AccessDenied if !current_user.has_authority_over?(User.find(params[:user_id]))
+    
     @users_skill=UsersSkill.find_by(user_id: params[:user_id], skill_id: params[:id])
+    
     @users_skill.update!(confirmed: true)
     render json: {status: :ok}
   end
